+++ src/Payment/Changelog.md
@@ -2,8 +2,8 @@
 
 ## 3.x Series
 
-## Unreleased
-##### 2023-XX-YY
+## 3.4.0
+##### 2023-01-25
 
 - Added `Configurable` interface to the `PaymentMethod` model
 - Changed the `PaymentMethod` model to use the `ConfigurableModel` trait